requirements.txt
alpha_vantage
kafka-python
pandas
backtesting
PyYAML
streamlit
plotly
tensorflow-cpu
scikit-learn
matplotlib
nltk
watchdog